Update: moved dep-agent package to use deployment-agent as main package
