+++ cmd/delete.go
